Add files via upload
MOSTLY FUNCTIONAL AS OF 4/4/21

+++ main.py
@@ -1,33 +1,90 @@
 #! python
 # main
 """TODO: 
-2. Fix file name to be more granular, add date
-3. QA/Test and error log
-4. Fix those errors
-5. Grid makers
+- QA/Test and error log:
+	7/8 is fucked
+- Triplets
+- Basslines/Hand Patterns = Alberti, Hannons, 
+- recycle repeats: so if sample < measure, the same note isn't repeated the whole page
 - CHORDS
 - Mixed rhythms
-6. Ryhtmic grid makers/Rests - basically Syncopation 2.0
-7. Reads XML and generates: random and grids
-8. Data Validation for all inputs, BLACKLISTS, WHITELISTS, REGEX'S, ERRORS
-9. GRAND STAFF - for piano sight reading, separate rhythms for bass/treble clef!!!!!OMG
-10. Create module called "tdoo", which collects and lists/updates in real time anything tagged tdoo, 
+- Ryhtmic grid makers/Rests - basically Syncopation 2.0
+- Reads XML and generates: random and grids
+- Data Validation for all inputs, BLACKLISTS, WHITELISTS, REGEX'S, ERRORS
+- Create module called "tdoo", which collects and lists/updates in real time anything tagged tdoo, 
 	it will have to parse comments - is that doable?
 
 
 """
 
-import notereader, config, xmlwriter, attributes, os, sys
+import notereader, config, xmlwriter, attributes, os, sys, time, gridmaker
 from pprint import pprint
 from pathlib import Path
 
 #MAIN PROGRAM LOOP
 while True:
-	notereader.run("1")
-
+	config.clear() #clearing all global attributes
+	program_select = input("1. Grid Maker or 2. Note Reader: ")
+	if program_select == "1":
+		gridmaker.run()
+	if program_select == "2":
+		notereader.run("1")
+	
+	#FILENAME VALIDATION
+	path = './xmlbounces'
+	while True:
+		if config.filename + '.xml' in os.listdir(path):
+			print()
+			filename_selection = input(f"WARNING: '{config.filename}.xml' already exists! Overwrite? (Y or N): ")
+			if filename_selection.lower() == 'n':
+				print()
+				newfilename = input("Enter New File Name: " + "\n")
+				if newfilename == config.filename:
+					print()
+					print("ERROR - file name already taken!")
+					continue
+				else:
+					config.filename = newfilename
+					break
+			if filename_selection.lower()== 'y':
+				break
+		else:
+			break
+	
 	attributes.setattr()
-
+	print()
+	print("ATTRIBUTES STAFF 1")
+	pprint(config.exercises_withattr_list)
+	print()
+	print("ATTRIBUTES STAFF 2")
+	pprint(config.exercises_withattr_list_staff2)
+	print()
+	print("MEASURE LENGTH STAFF 1: ")
+	print(config.measure_length_staff1)
+	print()
+	print("MEASURE LENGTH STAFF 2: ")
+	print(config.measure_length_staff2)
+	print()
+	print("CONFIG.NOTETYPE: ")
+	print(config.notetype)
+	print()
+	print("DIVS ")
+	print(config.divs)
+	print()
+	print("NOTETYPE_STAFF1 ")
+	print(config.notetype_staff1)
+	print()
+	print("NOTETYPE_STAFF2 ")
+	print(config.notetype_staff2)
+	print()
+	print("DUR_1 ")
+	print(config.test)
+	print()
+	print("DUR_2 ")
+	print(config.test2)
+	
 	xmlwriter.xmlwrite()
+	
 	#PROGRAM LOOP/PRINTING - TODO This loop may be wonky and/or redundant, double check later...
 	while True:
 			try:	
@@ -35,8 +92,15 @@ while True:
 				whitelist_run = ["r","ru","run","rn"]
 				whitelist_print = ["p","pr","pri","prin","print","prnt"]
 				if run.lower() in whitelist_run:
-					os.remove(f"xmlbounces/{config.filename}.xml")
-					break 
+					print()
+					delete = input(f"Delete file '{config.filename}.xml'? Y or N: ")
+					while True:
+						if delete.lower() == "y":
+							os.remove(f"xmlbounces/{config.filename}.xml")
+							break
+						if delete.lower() == "n":
+							break
+					break
 				if run.lower() in whitelist_print:
 					os.system('open '+ f'xmlbounces/{config.filename}.xml')
 					continue
@@ -45,6 +109,14 @@ while True:
 				else:
 					continue
 			except Exception:
+				print()
+				delete = input(f"Delete file '{config.filename}.xml'? Y or N: ")
+				while True:
+					if delete.lower() == "y":
+						os.remove(f"xmlbounces/{config.filename}.xml")
+						break
+					if delete.lower() == "n":
+						break
 				print("\n" + "---------- END ----------" + "\n")
 				sys.exit()
 	if run.lower() in whitelist_run:
